move CollapsablePrompts into its own component, and standardise interface spacing
