+++ src/styles/containers/_dyad-census.scss
@@ -57,7 +57,6 @@
     position: relative;
     flex: 1 auto;
     height: 100%;
-    margin-top: 1rem;
     min-height: 0;
   }
 
@@ -65,7 +64,7 @@
     max-height: 100%;
     background: var(--light-background);
     border-radius: var(--border-radius);
-    padding: unit(4);
+    padding: unit(2) unit(4);
     min-width: 50vw;
     text-align: center;
   }
