+++ src/styles/containers/_ego-form-interface.scss
@@ -1,5 +1,6 @@
 .ego-form {
   @include interface-centering(false);
+  padding: unit(4) 0;
   height: 100%;
   overflow: hidden;
   display: flex;
@@ -10,9 +11,9 @@
     width: 100%;
     border-radius: var(--border-radius);
     background: var(--form-intro-panel-background);
-    padding: unit(6);
+    padding: unit(2) unit(4);
     flex-direction: column;
-    margin: 3rem 0;
+    margin-bottom: unit(4);
     max-width: 55rem; // same as ego form max width
 
     h1 {
@@ -23,13 +24,15 @@
   &__form-container {
     height: 100%;
     width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
   }
 
   &__form-container-scroller {
-    height: 100%;
+    max-height: 100%;
     display: flex;
     flex-direction: column;
-    padding-bottom: unit(10);
 
     & > * {
       max-width: 55rem;
