+++ src/styles/containers/_interface.scss
@@ -13,7 +13,6 @@
     position: relative;
     flex: 1 auto;
     height: 100%;
-    margin-top: 1rem;
     min-height: 0;
   }
 }
