+++ src/styles/containers/_tie-strength-census.scss
@@ -52,7 +52,7 @@
     max-height: 100%;
     background: var(--light-background);
     border-radius: var(--border-radius);
-    padding: unit(4);
+    padding: unit(2) unit(4);
     min-width: 50vw;
     text-align: center;
   }
